% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCanadaInfo.R
\name{addCanadaInfo}
\alias{addCanadaInfo}
\title{addCanadaInfo.}
\usage{
addCanadaInfo(x)
}
\arguments{
\item{x}{dataset.}
}
\value{
Table with region info added
}
\description{
Adds Canada information to agData STATCAN datasets. Uses agData_STATCAN_Region_Table.
}
\examples{
xx <- agData_STATCAN_Crops \%>\% addCanadaInfo()
}
