% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRegionInfo.R
\name{addRegionInfo}
\alias{addRegionInfo}
\title{addRegionInfo}
\usage{
addRegionInfo(x)
}
\arguments{
\item{x}{dataset.}
}
\value{
Table with region info added
}
\description{
Adds region information to agData datasets. Uses both agData_FAO_Country_Table and agData_FAO_Region_Table.
}
\examples{
xx <- agData_FAO_Crops \%>\% addRegionInfo()
}
