% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_Barulina}
\alias{agData_Barulina}
\title{Lentil Barulina}
\source{
Barulina, Elena I. (1930) Lentils of the USSR and of other countries.
}
\description{
A dataset for lentil production from Barulina's Monograph.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurements (Unit)}{

\itemize{
  \item Area Harvested (Hectares)
  \item Production     (Tonnes)
  \item Yield          (kg/ha)
}
}

\keyword{datasets}
