% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Consumer_Price}
\alias{agData_FAO_Consumer_Price}
\title{FAO Consumer Price Indices}
\source{
http://www.fao.org/faostat/en/#data/CP
}
\description{
A dataset containing consumer price indices from FAOSTAT.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Months
  \item Unit
  \item Value
}
}

\section{Item}{

\itemize{
  \item Consumer Prices, Food Indices (2015 = 100)
  \item Consumer Prices, General Indices (2015 = 100)
  \item Food price inflation
}
}

\keyword{datasets}
