% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Country_Table}
\alias{agData_FAO_Country_Table}
\title{FAO Country Table}
\description{
A table of country information.
}
\section{agData_FAO_Country_Table}{

}

\section{Keys}{

\itemize{
  \item Country
  \item FAO_TABLE_NAME
  \item ISO2
  \item ISO3
  \item Lat
  \item Lon
  \item Region
  \item SubRegion
  \item DVDDVG
}
}

\keyword{datasets}
