% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Fertilizer}
\alias{agData_FAO_Fertilizer}
\title{FAO world population data}
\source{
http://www.fao.org/faostat/en/#data/RFN
}
\description{
FAO world population data
}
\section{Measurements}{

\itemize{
  \item Production
  \item Import Quantity
  \item Export Quantity
  \item Agricultural Use
  \item Prices Paid by Farmers
}
}

\keyword{datasets}
