% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Fertilizers}
\alias{agData_FAO_Fertilizers}
\title{FAO fertilizer data}
\source{
http://www.fao.org/faostat/en/#data/RFN + https://www.fao.org/faostat/en/#data/RFB + https://www.fao.org/faostat/en/#data/RA
}
\description{
FAO fertilizer data
}
\section{agData_FAO_Fertilizers}{

}

\section{Measurement}{

\itemize{
  \item Production
  \item Import Quantity
  \item Export Quantity
  \item Agricultural Use
  \item Prices Paid by Farmers
}
}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
