% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Food_Security2}
\alias{agData_FAO_Food_Security2}
\title{FAO Food Price Index #2}
\source{
https://www.fao.org/faostat/en/#data/FS
}
\description{
A dataset containing food price indices from FAOSTAT.
}
\section{Keys}{

\itemize{
  \item Date
  \item Year
  \item Month
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
