% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Producer_Prices}
\alias{agData_FAO_Producer_Prices}
\title{FAO Producer Price Indices}
\source{
http://www.fao.org/faostat/en/#data/PP
}
\description{
A dataset containing producer price indices from FAOSTAT.
}
\section{agData_FAO_Producer_Prices}{

}

\section{Measurement (Unit)}{

\itemize{
  \item Producer Price Index (2014-2016 = 100)
  \item Producer Price (USD/Tonne)
}
}

\section{Keys}{

\itemize{
  \item Area
  \item Item
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
