% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Region_Table}
\alias{agData_FAO_Region_Table}
\title{FAO Region Table}
\description{
A table of region information.
}
\section{agData_FAO_Region_Table}{

}

\section{Keys}{

\itemize{
  \item FAO_TABLE_NAME
  \item Region
  \item SubRegion
  \item Name
}
}

\keyword{datasets}
