% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Trade}
\alias{agData_FAO_Trade}
\title{FAO Trade Data
A dataset containing trade data from FAOSTAT.}
\source{
http://www.fao.org/faostat/en/#data/TP
}
\description{
FAO Trade Data
A dataset containing trade data from FAOSTAT.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Crop
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Export Quantity (tonnes)
  \item Import Quantity (tonnes)
  \item Export Value (1000 $USD)
  \item Import Value (1000 $USD)
}
}

\keyword{datasets}
