% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Trade_Value}
\alias{agData_FAO_Trade_Value}
\title{FAO Trade Data}
\source{
http://www.fao.org/faostat/en/#data/TP
}
\description{
A dataset containing trade data from FAOSTAT.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Export Value (USD)
  \item Import Value (USD)
}
}

\keyword{datasets}
