% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_ISAAA_Crop}
\alias{agData_ISAAA_Crop}
\title{ISAAA GE Crop Data}
\source{
https://www.isaaa.org/
}
\description{
A dataset from ISAAA for total GE crop area by Crop.
}
\section{agData_ISAAA_Crop}{

}

\section{Measurements}{

\itemize{
  \item Hectares
  \item Percent
  \item plus_minus
  \item Percent Change
}
}

\section{Keys}{

\itemize{
  \item Year
  \item Item
  \item Million.Hectares
  \item Percent
  \item plus_minus
  \item Percent.Change
}
}

\keyword{datasets}
