% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_ISAAA_Value}
\alias{agData_ISAAA_Value}
\title{ISAAA GE Crop Data}
\source{
https://www.isaaa.org/
}
\description{
A dataset from ISAAA for total value of GE crops.
}
\section{agData_ISAAA_Value}{

}

\section{Measurements}{

\itemize{
  \item Hectares
  \item Acres
}
}

\section{Keys}{

\itemize{
  \item Year
  \item Value
  \item Unit
}
}

\keyword{datasets}
