% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_NBT_Limits}
\alias{agData_NBT_Limits}
\title{NBT limitations
A dataset for survey data about limitation of NBT. Percentage of experts indicating that factor will be a barrier to NBT development.}
\source{
Lassoued R, Smyth SJ, Phillips PW & Hesseln H (2018) Regulatory uncertainty around new breeding techniques. Frontiers in plant science. 9.
}
\description{
NBT limitations
A dataset for survey data about limitation of NBT. Percentage of experts indicating that factor will be a barrier to NBT development.
}
\section{Measurements}{

\itemize{
  \item Hectares
  \item Acres
}
}

\keyword{datasets}
