% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_PeopleInAg}
\alias{agData_PeopleInAg}
\title{People in Ag}
\source{
?
}
\description{
A dataset from people in ag.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurements}{

\itemize{
  \item Total
  \item Percent
}
}

\keyword{datasets}
