% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_Consumer_Prices}
\alias{agData_STATCAN_Consumer_Prices}
\title{Statistics Canada Consumer Prices Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=1810000201
}
\description{
A dataset from Statistics Canada on Consumer Food Prices.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Month
  \item Product
  \item Value
}
}

\section{Table}{
 18-10-0002-01
}

\keyword{datasets}
