% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_Crops}
\alias{agData_STATCAN_Crops}
\title{Crops Statistics Canada Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210035901
}
\description{
A dataset from Statistics Canada on crop production.
}
\section{Table}{
 32-10-0359-01 (formerly CANSIM 001-0017)
}

\section{agData_STATCAN_Crops}{

}

\section{Measurement (Unit)}{

\itemize{
  \item Area Seeded      (Hectares)
  \item Area Harvested   (Hectares)
  \item Production       (Tonnes)
  \item Yield            (kg/ha)
  \item Total Farm Value (CAD)
}
}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
