% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_FarmLand_NoTill}
\alias{agData_STATCAN_FarmLand_NoTill}
\title{Statistics Canada Farmland Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210016201
}
\description{
A dataset from Statistics Canada on no till.
}
\section{Keys}{

}

\section{Measurement}{

\itemize{
  \item Commercial fertilizer
  \item Herbicides
  \item Insecticides
  \item Fungicides
  \item Total land prepared for seeding
  \item No-till seeding or zero-till seeding
  }
}

\section{Unit}{

\itemize{
  \item Number of farms reporting
  \item Hectares
}
}

\section{Table}{
 32-10-0162-01 (formerly CANSIM 004-0010)
}

\keyword{datasets}
