% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_FarmLand_Size}
\alias{agData_STATCAN_FarmLand_Size}
\title{Statistics Canada Farmland Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210015601
}
\description{
A dataset from Statistics Canada on farm land use.
}
\section{Table}{
 32-10-0156-01 (formerly CANSIM 004-0005)
}

\section{agData_STATCAN_FarmLand_Size}{

}

\section{Item (Unit)}{

\itemize{
  \item Total number of farms
  \item Under 10 acres
  \item 10 to 69 acres
  \item 70 to 129 acres
  \item 130 to 179 acres
  \item 180 to 239 acres
  \item 240 to 399 acres
  \item 400 to 559 acres
  \item 560 to 759 acres
  \item 760 to 1,119 acres
  \item 1,120 to 1,599 acres
  \item 1,600 to 2,239 acres
  \item 2,240 to 2,879 acres
  \item 2,880 to 3,519 acres
  \item 3,520 acres and above
}
}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
