% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_Population}
\alias{agData_STATCAN_Population}
\title{Statistics Canada population Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=1710000901
}
\description{
A dataset from Statistics Canada on population.
}
\section{Table}{
 17-10-0009-01 (formerly CANSIM 051-0005)
}

\section{agData_STATCAN_Population}{

}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Unit
  \item Value
}
}

\keyword{datasets}
