% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Population_AgeGender}
\alias{agData_STATCAN_Population_AgeGender}
\title{Statistics Canada population Data
A dataset from Statistics Canada on population by age and gender.}
\source{
https://www150.statcan.gc.ca/
}
\description{
Statistics Canada population Data
A dataset from Statistics Canada on population by age and gender.
}
\section{Keys}{

\itemize{
  \item Area
  \item Sex
  \item Year
  \item ...
}
}

\keyword{datasets}
