% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Population_Dynamics}
\alias{agData_STATCAN_Population_Dynamics}
\title{Statistics Canada population Data
A dataset from Statistics Canada on population dynamics.}
\source{
https://www150.statcan.gc.ca/
}
\description{
Statistics Canada population Data
A dataset from Statistics Canada on population dynamics.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement}{

\itemize{
  \item Births
  \item Deaths
  \item Immigrants
  \item Emigrants
}
}

\keyword{datasets}
