% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_UK_Yields}
\alias{agData_UK_Yields}
\title{Long term yield data for UK}
\source{
...
}
\description{
A dataset of long term yield data in UK.
}
\section{Keys}{

\itemize{
  \item Area
  \item Crop
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
