% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_USDA_Food_Security}
\alias{agData_USDA_Food_Security}
\title{USDA food security data}
\source{
https://www.ers.usda.gov/data-products/international-food-security/
}
\description{
A dataset from the USDA on food security.
}
\section{agData_USDA_Food_Security}{

}

\section{Keys}{

\itemize{
  \item Year
  \item Region
  \item SubRegion
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
