% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_USDA_GE_Crops}
\alias{agData_USDA_GE_Crops}
\title{USDA GE Crop Data}
\source{
https://www.ers.usda.gov/data-products/adoption-of-genetically-engineered-crops-in-the-us/
}
\description{
A dataset from the USDA on GE crop production.
}
\section{agData_USDA_GE_Crops}{

}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
