% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_USDA_Grains}
\alias{agData_USDA_Grains}
\title{USDA Crop Data}
\source{
https://www.nass.usda.gov/
}
\description{
A dataset from the USDA on crop production.
}
\section{agData_USDA_Grains}{

}

\section{Measurement (Unit)}{

\itemize{
  \item Production     (Tonnes)
  \item Area Seeded    (Hectares)
  \item Area Harvested (Hectares)
  \item Yield          (kg/ha)
}
}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
