% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_USDA_SugarBeet}
\alias{agData_USDA_SugarBeet}
\title{USDA Sugarbeet data}
\source{
https://www.nass.usda.gov/
}
\description{
A dataset from the USDA on sugarbeet production.
}
\section{agData_USDA_SugarBeet}{

}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
