% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_NA.R
\name{fill_NA}
\alias{fill_NA}
\title{fill_NA}
\usage{
fill_NA(x)
}
\arguments{
\item{x}{vector}
}
\value{
Vector with NA values replaced with the nearest above value
}
\description{
This function replaces NA values with the nearest above value.
}
\examples{
xx \%>\% mutate(Column1 = fill_NA(Column1))
}
