% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggBarRacer.R
\name{ggBarRacer}
\alias{ggBarRacer}
\title{Create a bar chart race animation.}
\usage{
ggBarRacer(
  df,
  cat_col,
  val_col,
  time_col,
  max_bars = 10,
  duration = 20,
  fps = 10,
  width = 1200,
  height = 900,
  title = "",
  renderer = gganimate::gifski_renderer()
)
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{cat_col}{The column containing the categories (for bars).}

\item{val_col}{The column containing the values (for bar heights).}

\item{time_col}{The column containing the time.}

\item{max_bars}{The maximum number of bars to show (defaults to 10).}

\item{duration}{The duration of the animation in seconds (defaults to 20).}

\item{fps}{The smoothness of the animation in frames per second (defaults to 10).}

\item{width}{The width of the animation in pixels (defaults to 1200).}

\item{height}{The height of the animation in pixels (defaults to 900).}

\item{title}{The title of the animation (defaults to blank).}

\item{renderer}{The renderer to use (defaults to gifski_renderer()).}
}
\value{
A bar chart race animation.
}
\description{
Adopted from https://github.com/jl5000/barRacer/blob/master/R/bar_chart_race.R
}
