% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvals.R
\name{mapvals}
\alias{mapvals}
\title{mapvals.}
\usage{
mapvals(x, from, to, warn_missing = TRUE)
}
\arguments{
\item{x}{list}

\item{from}{Entries to change}

\item{to}{Entries to change them to}

\item{warn_missing}{Logical, will warn if missing}
}
\value{
replaced list
}
\description{
ggplot theme.
}
