% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_Table.R
\name{pdf_Table}
\alias{pdf_Table}
\title{pdf_Table}
\usage{
pdf_Table(data, filename, rowsperpage = 36, width = 2.5, height = 11)
}
\arguments{
\item{data}{Data.}

\item{filename}{Name to give pdf file.}

\item{rowsperpage}{Rows per page.}

\item{width}{width of page.}

\item{height}{height of page.}
}
\value{
Table with mean and sd for trait.
}
\description{
This function calculates a pdf file for a table of data.
}
\examples{
pdf_Table(agData_FAO_Crops \%>\% filter(Area == "Canada", Crop == "Lentils") \%>\% select(Year, Measurement, Value), "myPDF.pdf", 41, 3.75, 12)
}
