% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_Info.R
\name{region_Info}
\alias{region_Info}
\title{region_Info}
\usage{
region_Info(x)
}
\arguments{
\item{x}{dataset.}
}
\value{
Table with region info added
}
\description{
Adds region information to agData datasets. Uses both agData_FAO_Country_Table and agData_FAO_Region_Table.
}
\examples{
xx <- agData_FAO_Crops \%>\% addRegionInfo()
}
