% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showmenorm.r
\name{showmenorm}
\alias{showmenorm}
\title{When you need a break...}
\usage{
showmenorm(x = NULL, list = FALSE)
}
\arguments{
\item{x}{the name or index of the video to display; if NULL, a random video is chosen}

\item{list}{a logical indicating if the list of available videos should be displayed}
}
\description{
Genetic data analysis can be a harsh, tiring, daunting task.
Sometimes, a mere break will not cut it.
Sometimes, you need Norm Macdonald.
}
\details{
Please send us more! Either pull request or submit an issue with a URL (use
\code{adegenetIssues()}).
}
\author{
Derek Michael Wright \email{derek.wright#usask.ca}
}
