% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_agData.R
\name{spread_agData}
\alias{spread_agData}
\title{spread_agData.}
\usage{
spread_agData(x)
}
\arguments{
\item{x}{dataset.}
}
\value{
Table
}
\description{
spreads data. #x <- spread(x, Element, Value)
}
\examples{
xx <- agData_FAO_Crops \%>\% agData_spread()
}
