% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData.R
\name{theme_agData}
\alias{theme_agData}
\title{theme_agData.}
\usage{
theme_agData(
  x,
  bgFill = "white",
  lineColor = "grey90",
  lineSize = 0.75,
  stripFill = "white",
  caption.adj = 1,
  caption.size = 7,
  ...
)
}
\arguments{
\item{x}{ggplot}

\item{bgFill}{background fill color}

\item{lineColor}{color of axis lines}

\item{lineSize}{border line size}

\item{stripFill}{strip background color}

\item{caption.adj}{hjust for the caption}

\item{caption.size}{size of the caption}
}
\value{
ggplot with agData theme
}
\description{
ggplot theme.
}
