% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_Barulina}
\alias{agData_Barulina}
\title{Lentil Barulina}
\source{
Barulina, Elena I. (1930) Lentils of the USSR and of other countries.
}
\description{
A dataset for lentil production from Barulina's Monograph.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Crop
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurements}{

\itemize{
  \item Area harvested
  \item Production
  \item Yield
}
}

\keyword{datasets}
