% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Country_Table}
\alias{agData_FAO_Country_Table}
\title{FAO Country Table
A table of country info from FAOSTAT.}
\source{
http://www.fao.org/faostat/
}
\description{
FAO Country Table
A table of country info from FAOSTAT.
}
\section{Helpful Codes}{

\itemize{
  \item unique(agData_FAO_Country_Table$Name)
  \item unique(agData_FAO_Country_Table$Region)
  \item unique(agData_FAO_Country_Table$SubRegion)
}
}

\keyword{datasets}
