% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Country_Table}
\alias{agData_FAO_Country_Table}
\title{FAO Country Table
A table of country info from FAOSTAT.}
\source{
http://www.fao.org/faostat/
}
\description{
FAO Country Table
A table of country info from FAOSTAT.
}
\keyword{datasets}
