% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Crops}
\alias{agData_FAO_Crops}
\title{FAO Crop Production Data
A dataset containing crop data from FAOSTAT.}
\source{
http://www.fao.org/faostat/
}
\description{
FAO Crop Production Data
A dataset containing crop data from FAOSTAT.
}
\section{Helpful Codes}{

\itemize{
  \item levels(agData_FAO_Crops$Area)
  \item levels(agData_FAO_Crops$Item)
  \item addRegionInfo(agData_FAO_Crops)
}
}

\section{Elements}{

\itemize{
  \item Production     (tonnes)
  \item Area harvested (ha)
  \item Yield          (tonnes/ha)
}
}

\keyword{datasets}
