% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Crops2}
\alias{agData_FAO_Crops2}
\title{FAO Crop Production Data
A dataset containing crop aggregate data from FAOSTAT.}
\source{
http://www.fao.org/faostat/en/#data/QC
}
\description{
FAO Crop Production Data
A dataset containing crop aggregate data from FAOSTAT.
}
\section{Measurement (Unit)}{

\itemize{
  \item Production     (tonnes)
  \item Area harvested (hectares)
  \item Yield          (t/ha)
}
}

\keyword{datasets}
