% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Crops3}
\alias{agData_FAO_Crops3}
\title{FAO Crop Production Data}
\source{
http://www.fao.org/faostat/en/#data/QC
}
\description{
A dataset containing crop data from FAOSTAT.
}
\section{agData_FAO_Crops3}{
NA
}

\section{Measurement (Unit)}{

\itemize{
  \item Production     (Tonnes)
  \item Area Harvested (Hectares)
  \item Yield          (kg/ha)
}
}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
