% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Food_Prices}
\alias{agData_FAO_Food_Prices}
\title{FAO Food Price Index}
\source{
https://www.fao.org/worldfoodsituation/foodpricesindex/en/
}
\description{
A dataset containing food price indices from FAOSTAT.
}
\section{Keys}{

\itemize{
  \item Date
  \item Year
  \item Month
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
