% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_LandUse}
\alias{agData_FAO_LandUse}
\title{FAO Land Use Data
A dataset containing land use data from FAOSTAT.}
\source{
http://www.fao.org/faostat/
}
\description{
FAO Land Use Data
A dataset containing land use data from FAOSTAT.
}
\section{Items}{

\itemize{
  \item levels(agData_FAO_LandUse$Area)
  \item levels(agData_FAO_LandUse$Item)
  \item addRegionInfo(agData_FAO_LandUse)
}
}

\section{Elements}{

\itemize{
  \item Area                           (1000 ha)
  \item Carbon stock in living biomass (million tonnes)
}
}

\keyword{datasets}
