% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_LandUse}
\alias{agData_FAO_LandUse}
\title{FAO Land Use Data}
\source{
http://www.fao.org/faostat/en/#data/RL
}
\description{
A dataset containing land use data from FAOSTAT.
}
\section{agData_FAO_LandUse}{
NA
}

\section{Item}{

\itemize{
  \item Country area
  \item Land area
  \item Agricultural area
  \item ...
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Area                           (1000 hectares)
  \item Carbon stock in living biomass (million tonnes)
}
}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
