% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_FAO.R
\docType{data}
\name{agData_FAO_Livestock}
\alias{agData_FAO_Livestock}
\title{FAO Livestock Data}
\source{
http://www.fao.org/faostat/en/#data/QA
}
\description{
A dataset containing livestock data from FAOSTAT.
}
\section{agData_FAO_Livestock}{
NA
}

\section{Measurement (Unit)}{

\itemize{
  \item Stocks (head or 1000 head)
}
}

\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
