% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Livestock}
\alias{agData_FAO_Livestock}
\title{FAO Livestock Data
A dataset containing livestock data from FAOSTAT.}
\source{
http://www.fao.org/faostat/
}
\description{
FAO Livestock Data
A dataset containing livestock data from FAOSTAT.
}
\section{Helpful Codes}{

\itemize{
  \item levels(agData_FAO_Livestock$Area)
  \item levels(agData_FAO_Livestock$Item)
  \item addRegionInfo(agData_FAO_Livestock)
}
}

\section{Elements}{

\itemize{
  \item Stocks (head or 1000 head)
}
}

\keyword{datasets}
