% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Population}
\alias{agData_FAO_Population}
\title{FAO world population data}
\source{
http://www.fao.org/faostat/en/#data/OA
}
\description{
FAO world population data
}
\section{Measurements}{

\itemize{
  \item Total
  \item Male
  \item Female
  \item Rural
  \item Urban
}
}

\keyword{datasets}
