% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_FAO_Region_Table}
\alias{agData_FAO_Region_Table}
\title{FAO Region Table
A table of Region info from FAOSTAT.}
\source{
http://www.fao.org/faostat/
}
\description{
FAO Region Table
A table of Region info from FAOSTAT.
}
\section{Helpful Codes}{

\itemize{
  \item unique(agData_FAO_Region_Table$Region)
  \item unique(agData_FAO_Region_Table$SubRegion)
}
}

\keyword{datasets}
