% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_ISAAA_Area}
\alias{agData_ISAAA_Area}
\title{ISAAA GE Crop Data}
\source{
https://www.isaaa.org/
}
\description{
A dataset from ISAAA for total GE crop area.
Note: this data set includes cultivation of GE tobacco in China for 1996 and 1997.
}
\section{agData_ISAAA_Area}{
NA
}

\section{Measurements}{

\itemize{
  \item Hectares
  \item Acres
}
}

\section{Keys}{

\itemize{
  \item Year
  \item Hectares
  \item plus_minus
  \item Percent.Change
}
}

\keyword{datasets}
