% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_ISAAA_Country}
\alias{agData_ISAAA_Country}
\title{ISAAA GE Country Data}
\source{
https://www.isaaa.org/
}
\description{
A dataset from ISAAA for total GE crop area by Country.
Note: data is rounded to the nearest 100,000 hectares.
}
\section{Keys}{

\itemize{
  \item Year
  \item Crop
  \item Million.Hectares
  \item Percent
  \item plus_minus
  \item Percent.Change
}
}

\section{Measurements}{

\itemize{
  \item Hectares
  \item Percent
  \item plus_minus
  \item Percent Change
}
}

\keyword{datasets}
