% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_ISAAA_Crop}
\alias{agData_ISAAA_Crop}
\title{ISAAA GE Crop Data
A dataset from ISAAA for total GE crop area by Crop.}
\source{
https://www.isaaa.org/
}
\description{
ISAAA GE Crop Data
A dataset from ISAAA for total GE crop area by Crop.
}
\section{Measurements}{

\itemize{
  \item Hectares
  \item Percent
  \item plus_minus
  \item Percent Change
}
}

\keyword{datasets}
