% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_ISAAA_CropPercent}
\alias{agData_ISAAA_CropPercent}
\title{ISAAA GE Crop Data}
\source{
https://www.isaaa.org/
}
\description{
A dataset from ISAAA for total GE crop area by Crop percent.
}
\section{Keys}{

\itemize{
  \item Year
  \item Item
  \item Global.Area
  \item Transgenic.Crop.Area
  \item Percent.Transgenic
}
}

\section{Measurements}{

\itemize{
  \item Hectares
  \item Acres
}
}

\keyword{datasets}
