% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_ISAAA_CropPercent}
\alias{agData_ISAAA_CropPercent}
\title{ISAAA GE Crop Data
A dataset from ISAAA for total GE crop area by Crop percent.}
\source{
https://www.isaaa.org/
}
\description{
ISAAA GE Crop Data
A dataset from ISAAA for total GE crop area by Crop percent.
}
\section{Measurements}{

\itemize{
  \item Hectares
  \item Acres
}
}

\keyword{datasets}
