% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_ISAAA_DVDDVG}
\alias{agData_ISAAA_DVDDVG}
\title{ISAAA GE Crop Data}
\source{
https://www.isaaa.org/
}
\description{
A dataset from ISAAA for total GE crop area in industrialized and devloping nations.
Note: this data set includes cultivation of GE tobacco in China for 1996 and 1997.
}
\section{agData_ISAAA_DVDDVG}{
NA
}

\section{Measurements}{

\itemize{
  \item Hectares
  \item Percent
}
}

\section{Keys}{

\itemize{
  \item Year
  \item Area
  \item Hectares
  \item Percent
}
}

\keyword{datasets}
