% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_MaizeContest}
\alias{agData_MaizeContest}
\title{Maize Contest}
\source{
https://www.ncga.com/
}
\description{
A dataset for Maize Yield Contest.
}
\section{Keys}{

\itemize{
  \item Item
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurements}{

\itemize{
  \item Irrigated
  \item Non-Irrigated
}
}

\keyword{datasets}
