% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_MaizeContest}
\alias{agData_MaizeContest}
\title{Maize Contest
A dataset for Maize Yield Contest.}
\source{
Corn Growers Association.
}
\description{
Maize Contest
A dataset for Maize Yield Contest.
}
\section{Keys}{

\itemize{
  \item Crop
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurements}{

\itemize{
  \item Irrigated
  \item Non-Irrigated
}
}

\keyword{datasets}
