% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_PopFert}
\alias{agData_PopFert}
\title{Population with and without fertilizers}
\source{
https://www.nature.com/articles/ngeo325
}
\description{
A dataset of global population with and without fertilizers.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\keyword{datasets}
