% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_SKAG_Lentil_Exports}
\alias{agData_SKAG_Lentil_Exports}
\title{Lentil Export Destinations}
\source{
https://publications.saskatchewan.ca/#/categories/4426
}
\description{
A dataset from SK AG for the top 20 Lentil export countries each year.
}
\section{Keys}{

\itemize{
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Exports (Tonnes)
}
}

\keyword{datasets}
