% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_SKAG_Specialty_Crops}
\alias{agData_SKAG_Specialty_Crops}
\title{Sask AG Specialty Crop Data}
\source{
https://publications.saskatchewan.ca/#/categories/4426
}
\description{
A dataset from Sask Ag on specialty crop production.
}
\section{Keys}{

\itemize{
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Production     (Tonnes)
  \item Area Seeded    (Hectares)
  \item Area Harvested (Hectares)
  \item Yield          (kg/ha)
}
}

\keyword{datasets}
