% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_Beehives}
\alias{agData_STATCAN_Beehives}
\title{Statistics Canada - HoneyBee Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210035301
}
\description{
A dataset from Statistics Canada on honeybee production.
}
\section{Keys}{

\itemize{
  \item Area
  \item Item
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Beekeepers           (Thousand)
  \item Colonies             (Thousand)
  \item Colonies / Beekeeper (Number)
  \item Production           (Tonnes)
  \item Yield                (kg/colony)
  \item Value of Honey       (Million CAD)
  \item Value / Colony       (CAD/colony)
}
}

\section{Table}{
 32-10-0353-01 (formerly CANSIM 001-0007)
}

\keyword{datasets}
