% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Beehives}
\alias{agData_STATCAN_Beehives}
\title{Statistics Canada - HoneyBee Data
A dataset from Statistics Canada on honeybee production.}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=3210035301
}
\description{
Statistics Canada - HoneyBee Data
A dataset from Statistics Canada on honeybee production.
}
\section{Keys}{

\itemize{
  \item Area
  \item Crop
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Beekeepers         (number)
  \item Colonies           (number)
  \item Colonies/Beekeeper (number)
  \item Production         (tonnes)
  \item Yield              (kg/colony)
  \item Value              (million $CAD)
}
}

\section{Table}{
 32-10-0353-01 (formerly CANSIM 001-0007)
}

\keyword{datasets}
