% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Crops}
\alias{agData_STATCAN_Crops}
\title{Crops Statistics Canada Data
A dataset from Statistics Canada on crop production.}
\source{
https://www150.statcan.gc.ca/
}
\description{
Crops Statistics Canada Data
A dataset from Statistics Canada on crop production.
}
\section{Helpful Codes}{

\itemize{
  \item levels(agData_STATCAN_Crops$Area)
  \item levels(agData_STATCAN_Crops$Item)
  \item add_Canada_Info(agData_STATCAN_Crops)
}
}

\section{Elements}{

\itemize{
  \item Production         (tonnes)
  \item Area seeded        (ha)
  \item Area harvested     (ha)
  \item Yield              (tonnes/ha)
  \item Average farm price (dollars per tonne)
}
}

\section{Table}{
 ?
}

\keyword{datasets}
