% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Crops2}
\alias{agData_STATCAN_Crops2}
\title{Crops Statistics Canada Data
A dataset from Statistics Canada on crop aggregate data.}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210035901
}
\description{
Crops Statistics Canada Data
A dataset from Statistics Canada on crop aggregate data.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Crop
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Production         (tonnes)
  \item Area seeded        (hectares)
  \item Area harvested     (hectares)
  \item Yield              (tonnes/ha)
  \item Average farm price (dollars per tonne)
}
}

\section{Table}{
 32-10-0359-01 (formerly CANSIM 001-0017)
}

\keyword{datasets}
