% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_FarmLand_Crops}
\alias{agData_STATCAN_FarmLand_Crops}
\title{Statistics Canada Farmland Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210015401
}
\description{
A dataset from Statistics Canada on farm land use.
}
\section{Keys}{

}

\section{Measurement}{

\itemize{
  \item Number of farms reporting
  \item Hectares
}
}

\section{Table}{
 32-10-0154-01 (formerly CANSIM 004-0003)
}

\keyword{datasets}
