% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_FarmLand_Crops}
\alias{agData_STATCAN_FarmLand_Crops}
\title{Statistics Canada Farmland Data
A dataset from Statistics Canada on farm land use.}
\source{
https://www150.statcan.gc.ca/
}
\description{
Statistics Canada Farmland Data
A dataset from Statistics Canada on farm land use.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Item
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Area (Number of farms reporting)
  \item Area (Acres)
  \item Area (Hectares)
  \item Area (Average area in acres per farm reporting)
}
}

\keyword{datasets}
