% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_FarmLand_Farms}
\alias{agData_STATCAN_FarmLand_Farms}
\title{Statistics Canada Farmland Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210015201
}
\description{
A dataset from Statistics Canada on farm land use.
}
\section{Table}{
 32-10-0152-01 (formerly CANSIM 004-0001)
}

\section{agData_STATCAN_FarmLand_Farms}{
NA
}

\section{Measurement}{

\itemize{
  \item Total number of farms
  \item Total area of farms
  \item Area owned
  \item Area rented or leased from others
}
}

\section{Keys}{

}

\keyword{datasets}
