% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_FarmLand_Prices}
\alias{agData_STATCAN_FarmLand_Prices}
\title{Statistics Canada Farmland prices}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210004701
}
\description{
A dataset from Statistics Canada on farm land prices.
}
\section{Table}{
 32-10-0047-01 (formerly CANSIM 002-0003)
}

\section{agData_STATCAN_FarmLand_Prices}{
NA
}

\section{Keys}{

}

\section{Area}{

}

\keyword{datasets}
