% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_FarmReceipts}
\alias{agData_STATCAN_FarmReceipts}
\title{Statistics Canada Farm Receipt Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210004501
}
\description{
A dataset from Statistics Canada on farm receipts.
}
\section{NA}{
 32-10-0045-01
}

\section{Table}{
 32-10-0045-01
}

\keyword{datasets}
