% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Livestock}
\alias{agData_STATCAN_Livestock}
\title{Statistics Canada Livestock Data
A dataset from Statistics Canada on livestock production.}
\source{
https://www150.statcan.gc.ca/
}
\description{
Statistics Canada Livestock Data
A dataset from Statistics Canada on livestock production.
}
\section{Helpful Codes}{

\itemize{
  \item levels(agData_STATCAN_Livestock$Area)
  \item levels(agData_STATCAN_Livestock$Item)
  \item add_Canada_Info(agData_STATCAN_Livestock)
}
}

\section{Element}{

\itemize{
  \item Average number of animals (Number)
  \item Number of animals         (Number)
  \item Number of farms reporting (Number)
}
}

\section{Table}{
 ?
}

\keyword{datasets}
