% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Livestock}
\alias{agData_STATCAN_Livestock}
\title{Statistics Canada Livestock Data
A dataset from Statistics Canada on livestock production.}
\source{
https://www150.statcan.gc.ca/
}
\description{
Statistics Canada Livestock Data
A dataset from Statistics Canada on livestock production.
}
\section{Keys}{

\itemize{
  \item Area
  \item Animal
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Average number of animals (number)
  \item Number of animals         (number)
  \item Number of farms reporting (number)
}
}

\keyword{datasets}
