% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_Livestock}
\alias{agData_STATCAN_Livestock}
\title{Statistics Canada Livestock Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=3210015501
}
\description{
A dataset from Statistics Canada on livestock production.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Animal
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Average number of animals (number)
  \item Number of animals         (number)
  \item Number of farms reporting (number)
}
}

\section{Table}{
 32-10-0155-01 (formerly CANSIM 004-0004)
}

\keyword{datasets}
