% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Population}
\alias{agData_STATCAN_Population}
\title{Statistics Canada population Data
A dataset from Statistics Canada on population.}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=1710000901&cubeTimeFrame.startMonth=01&cubeTimeFrame.startYear=1946&cubeTimeFrame.endMonth=10&cubeTimeFrame.endYear=2020&referencePeriods=19460101%2C20201001
}
\description{
Statistics Canada population Data
A dataset from Statistics Canada on population.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Unit
  \item Value
}
}

\section{Table}{
 17-10-0009-01 (formerly CANSIM 051-0005)
}

\keyword{datasets}
