% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Population}
\alias{agData_STATCAN_Population}
\title{Statistics Canada population Data
A dataset from Statistics Canada on population.}
\source{
https://www150.statcan.gc.ca/
}
\description{
Statistics Canada population Data
A dataset from Statistics Canada on population.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Unit
  \item Value
}
}

\keyword{datasets}
