% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_STATCAN.R
\docType{data}
\name{agData_STATCAN_Population_AgeGender}
\alias{agData_STATCAN_Population_AgeGender}
\title{Statistics Canada population Data}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/tv.action?pid=1710000501
}
\description{
A dataset from Statistics Canada on population by age and gender.
}
\section{Keys}{

\itemize{
  \item Area
  \item Sex
  \item Year
  \item ...
}
}

\section{Table}{
 17-10-0005-01 (formerly CANSIM 051-0001)
}

\keyword{datasets}
