% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Population_Dynamics}
\alias{agData_STATCAN_Population_Dynamics}
\title{Statistics Canada population Data
A dataset from Statistics Canada on population dynamics.}
\source{
https://www150.statcan.gc.ca/t1/tbl1/en/cv.action?pid=1710000801
}
\description{
Statistics Canada population Data
A dataset from Statistics Canada on population dynamics.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement}{

\itemize{
  \item Births
  \item Deaths
  \item Immigrants
  \item Emigrants
}
}

\section{Table}{
 17-10-0008-01 (formerly CANSIM 051-0004)
}

\keyword{datasets}
