% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_STATCAN_Region_Table}
\alias{agData_STATCAN_Region_Table}
\title{Statistics Canada - Region Info
A dataset of Region info for Canada.}
\description{
Statistics Canada - Region Info
A dataset of Region info for Canada.
}
\section{Keys}{

\itemize{
  \item Area
  \item Lat
  \item Lon
}
}

\keyword{datasets}
