% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_USDA_Crops}
\alias{agData_USDA_Crops}
\title{USDA Crop Data
A dataset from the USDA on crop production.}
\source{
https://www.nass.usda.gov/
}
\description{
USDA Crop Data
A dataset from the USDA on crop production.
}
\section{Item}{

\itemize{
  \item Barley
  \item Maize
  \item Oats
  \item Sorghum
  \item Wheat
}
}

\section{Element}{

\itemize{
  \item Production     (tonnes)
  \item Area seeded    (ha)
  \item Area harvested (ha)
  \item Yield          (tonnes/ha)
}
}

\keyword{datasets}
