% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDocumentation.R
\docType{data}
\name{agData_USDA_Crops}
\alias{agData_USDA_Crops}
\title{USDA Crop Data
A dataset from the USDA on crop production.}
\source{
https://www.nass.usda.gov/
}
\description{
USDA Crop Data
A dataset from the USDA on crop production.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Crop
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item Production     (tonnes)
  \item Area seeded    (hectares)
  \item Area harvested (hectares)
  \item Yield          (t/ha)
}
}

\keyword{datasets}
