% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets_Other.R
\docType{data}
\name{agData_USDA_TFP}
\alias{agData_USDA_TFP}
\title{USDA Agricultureal Productivity data}
\source{
https://www.nass.usda.gov/
}
\description{
A dataset from the USDA on agricultural productivity production.
}
\section{Keys}{

\itemize{
  \item Area
  \item Year
  \item Measurement
  \item Unit
  \item Value
}
}

\section{Measurement (Unit)}{

\itemize{
  \item TFP
  \item Inputs
  \item Outputs
  \item ...
}
}

\keyword{datasets}
