% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData.R
\name{theme_agData}
\alias{theme_agData}
\title{theme_agData.}
\usage{
theme_agData(
  x,
  linesize = 0.75,
  bgFill = "grey95",
  stripFill = "white",
  lineColor = "white",
  rotx = F,
  ...
)
}
\arguments{
\item{x}{ggplot}

\item{linesize}{border line size}

\item{bgFill}{background fill color}

\item{stripFill}{strip background color}

\item{lineColor}{color of axis lines}
}
\value{
ggplot
}
\description{
ggplot theme.
}
