% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData.R
\name{theme_agData}
\alias{theme_agData}
\title{theme_agData.}
\usage{
theme_agData(x)
}
\arguments{
\item{x}{ggplot}
}
\value{
ggplot
}
\description{
ggplot theme.
}
\examples{
xx <- agData_FAO_Crops \%>\% agData_spread()
}
