% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData2.R
\name{theme_agData2}
\alias{theme_agData2}
\title{theme_agData2.}
\usage{
theme_agData2(
  x,
  bgFill = "grey95",
  lineColor = "white",
  lineSize = 0.75,
  stripFill = "white",
  caption.adj = 1,
  ...
)
}
\arguments{
\item{x}{ggplot}

\item{bgFill}{background fill color}

\item{lineColor}{color of axis lines}

\item{lineSize}{border line size}

\item{stripFill}{strip background color}

\item{caption.adj}{hjust for the caption}
}
\value{
ggplot with agData theme
}
\description{
ggplot theme.
}
