% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData2.R
\name{theme_agData2}
\alias{theme_agData2}
\title{theme_agData.}
\usage{
theme_agData2(
  x,
  bgFill = "white",
  lineColor = "grey80",
  linesize = 0.75,
  stripFill = "white",
  ...
)
}
\arguments{
\item{x}{ggplot}

\item{bgFill}{background fill color}

\item{lineColor}{color of axis lines}

\item{linesize}{border line size}

\item{stripFill}{strip background color}
}
\value{
ggplot with agData theme
}
\description{
ggplot theme.
}
