% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData_col.R
\name{theme_agData_col}
\alias{theme_agData_col}
\title{theme_agData_col.}
\usage{
theme_agData_col(
  x,
  bgFill = "white",
  lineColor = "grey90",
  linesize = 0.75,
  stripFill = "white",
  horizontal = F,
  ...
)
}
\arguments{
\item{x}{ggplot}

\item{bgFill}{background fill color}

\item{lineColor}{color of axis lines}

\item{linesize}{border line size}

\item{stripFill}{strip background color}

\item{horizontal}{Logical, use if you want the y axis gridlines removed}
}
\value{
ggplot with agData theme
}
\description{
ggplot theme.
}
