% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData_pie.R
\name{theme_agData_pie}
\alias{theme_agData_pie}
\title{theme_agData_pie.}
\usage{
theme_agData_pie(x, ...)
}
\arguments{
\item{x}{ggplot}
}
\value{
ggplot with agData theme for pie graphs
}
\description{
ggplot theme.
}
