% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData_pie_blank.R
\name{theme_agData_pie_blank}
\alias{theme_agData_pie_blank}
\title{theme_agData_pie_blank.}
\usage{
theme_agData_pie_blank(x, caption.adj = 1, ...)
}
\arguments{
\item{x}{ggplot}

\item{caption.adj}{hjust for the caption}
}
\value{
ggplot with agData theme for pie graphs
}
\description{
ggplot theme.
}
