% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_agData_pie_blank.R
\name{theme_agData_pie_blank}
\alias{theme_agData_pie_blank}
\title{theme_agData_pie_blank.}
\usage{
theme_agData_pie_blank(
  x,
  bgFill = "grey95",
  lineColor = "white",
  linesize = 0.75,
  stripFill = "white",
  ...
)
}
\arguments{
\item{x}{ggplot}
}
\value{
ggplot with agData theme for pie graphs
}
\description{
ggplot theme.
}
